Export vae_encoder, upgrade img2img, add test
