+++ src/diffusers/__init__.py
@@ -58,8 +58,11 @@ else:
     from .utils.dummy_torch_and_transformers_objects import *  # noqa F403
 
 if is_torch_available() and is_transformers_available() and is_onnx_available():
-    from .pipelines import StableDiffusionOnnxPipeline
-    from .pipelines import StableDiffusionInpaintOnnxPipeline
+    from .pipelines import (
+        StableDiffusionImg2ImgOnnxPipeline,
+        StableDiffusionInpaintOnnxPipeline,
+        StableDiffusionOnnxPipeline,
+    )
 else:
     from .utils.dummy_torch_and_transformers_and_onnx_objects import *  # noqa F403
